requirements.txt
aniso8601==9.0.1
apispec==6.6.1
blinker==1.8.1
click==8.1.7
Cython==3.0.10
Flask==2.3.3
flask-apispec==0.11.4
Flask-RESTful==0.3.10
gunicorn==21.2.0
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
madmom==0.16.1
MarkupSafe==2.1.5
marshmallow==3.21.1
numpy==1.23.5
packaging==24.0
PyJWT==2.8.0
pytz==2024.1
six==1.16.0
webargs==8.4.0
Werkzeug==3.0.2
zipp==3.18.1
